requirements.txt
aiogram~=3.23.0
gspread~=6.2.1
python-dotenv==1.0.1
oauth2client==4.1.3